Update transformer.md
Multi-head attention is explained using W_o explicitly, but the implementation of class MultiHeadAttention does not make the dense layer for W_o. I think that it is because W_o is d_o by d_o. Implicitly, the code assumes W_o is the identity layer. 
It would be better either if you explain why W_o is not shown in the implementation of the class or if you make another dense layer for W_o in the class.