Update anchor.md
I double checked the return values of the function MultiBoxPrior. I noticed that if [a, b, c, d] is for an anchor box returned by the function, then (c - a) * w which is the width of the anchor box  is not w * s * sqrt(r) but delta * w * s * sqrt(r), where delta is a constant and about 0.7706. The height of the anchor box is (d - b) * h which is equal to h * s / sqrt(r) as expected.
It would be nicer if you explain such a difference.