Change license to CC-BY-SA-4.0, add MIT-0 license for sample code, add code of conduct
This commit changes the license of the text of this book to
CC-BY-SA-4.0, changes the license of any included sample or reference
code to MIT-0, and adds the Amazon Code of Conduct.

All past commits to this repo have been reviewed and audited, and were
either authored by Amazon employees with copyright owned by Amazon, or
were authored by external contributors who have agreed to have their
contributon relicensed to these licenses.

Signed-off-by: Mark Atwood <atwoodm@amazon.com>