add PULL_REQUEST_TEMPLATE that contains contribution agreement
This commit adds a PULL_REQUEST_TEMPLATE that contains the statement
"By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice."

Pull requests without this statement will not be accepted.

This allows the book to be relicensed as needed in the future.

Signed-off-by: Mark Atwood <atwoodm@amazon.com>