[MRG] Correct typos (#522)
The intention of this PR is to fix two typos. 

However, I want to point out that the sentence `And if our classifier is a state of the art deep neural network.` seems incomplete to me. Although It might be your intention to keep open like this to let the reader infer what you mean.