backprop.md: Make vector y boldface.
Given that the output $\mathbf{o}$ is a vector, it seems likely that the label $y$ would be a vector, too.  (Unless, of course, the loss function is converting a scalar label into a one-shot label -- but in that case, I would expect a softmax output layer.)

I hope I can update my proposed change because I also need to edit the figure.