Update softmax-regression-concise.md (#1593)
* Update softmax-regression-concise.md

The current version of the modified softmax function and text (equation 3.7.1)  is not clear. It is not clear how it helps with the overflow error that can happen explained earlier in the paragraph; even though this equation helps with not evaluating 𝑜𝑗 in the numerator, it still has to be evaluated in the sum taken in -log(∑𝑘exp(𝑜𝑘)).  It looks like that this section is trying to show Log-sum-exp trick explained here: https://timvieira.github.io/blog/post/2014/02/11/exp-normalize-trick/ (and briefly mentioned in the text), so that equation 3.7.1 should include the normalization.  I included the normalization in the proposed change, and text more clearly explaining what it is doing.

* Update softmax-regression-concise.md

fix grammer issue