Include get_params in d2l tensorflow
If the code for 8.5 rnn-scratch is supposed to work if we load the "saved" code from the d2l library, we need the get_params method called by train_ch8.
Unfortunately, this method is not saved in. However, it should.
To test the code, change train_ch8 to d2l.train_ch8 in e.g. rnn-scratch.ipynb (tensorflow version) and you will get an error.