Add get_params as parameter for train_ch8
If we want to save train_ch8 in tensorflow.py, we need access to get_params.
Since this method is not saved and will be different for e.g. lstm, we need to pass it in as a parameter.
Changing train_ch8 to d2l.train_ch8 will currently cause an error!