[TF]Add get_params as a parameter for train_ch8 for tensorflow (#1672)
* Include get_params as argument for train_ch8 in tensorflow.py

train_ch8 needs get_params but this method does not exist in tensorflow.py.
We should probably pass it in as an argument since this function will be different for e.g. lstm.
change train_ch8 to d2l.train_ch8 in rnn-scratch.ipynb to see the problem.

* Add get_params as parameter for train_ch8

If we want to save train_ch8 in tensorflow.py, we need access to get_params.
Since this method is not saved and will be different for e.g. lstm, we need to pass it in as a parameter.
Changing train_ch8 to d2l.train_ch8 will currently cause an error!