Change traning loop of TF for chapter 8
The parameters of the model should be stored in the model.
They are initialized during the first model call, otherwise trainable_variables returns an empty list.
Therefore, they are now accessed after the first call and not before.
The cratch model uses trainable_variables to store the parameters which makes the code compatible
with both the scratch as well as the concise Keras implementation.