review number of flops operation per sec (#2260)
* review number of flops operation per sec

The multiplication of two matrices with an identical form of (256,256) implies 256^3 = 16777216 multiplications of real numbers which is approximately 0.02*10^9. So unless I have omitted something in my reasoning, the 2 must be replaced by 0.02 while keeping the rounding to simplify.

* Multiplication and Addition taken into account

As discussed with cheungdaven, in order to follow the logic initiated by the comments preceding the introduction of gigaflops in the Minibatch Stochastic Gradient Descent (12.5) section, the number of multiplications (approximately 0.02*10^9) AND the number of additions (approximately 0.02*10^9) must be taken into account.