ViTBlock implementation match schema
Since the schema presented at the top of the section is in agreement with the schema presented in the original paper (https://arxiv.org/abs/2010.11929), I therefore take this schema as the first reference to match the implementation with it. More specifically, I modify the implementation of the forward function of the ViTBlock object that I consider not to match the schema.