JAX: Refactor dropout to handle training attr
An alternate potentially better method could be to make
`training` an argument to the __call__/forward method of
the class and then pass it around the trainer setting the
boolean value to True for training and False for validation.

See https://jax.readthedocs.io/en/latest/faq.html#how-to-use-jit-with-methods
for the different strategies to mark class methods with JIT.
Currently we use strategy 2 which is simple/readable but may
lead to unexpected results if self itself changes some attrs.

When jitting the code, carrying around state is a little shady.
This can be looked into in the future.

JAX: Add missing partial import

JAX: Add missing optax import