JAX: Support multiple inputs X for example Seq2Seq data
Earlier the tuple unpacking took place when calling the
loss function in the training step. This wasn't elegant,
since we need to mark some args as static while using jax.jit
for the loss. If the X argument i.e. tuple contains more
than one input like in Machine Translation Dataset the static_argnums
values will be disrupted and will be wrong. Hence we delay the unpacking
of the input batch tuple to when the fwd pass is called by `apply_fn`.