Update W matrix subscript to d
The weight matrix is said to be an element of $\mathbb{R}^{d \times h}$, but is denoted as $\mathbf{W}_{xh}$ with the first subscript $x$ throughout the page. $x$ is nowhere defined, so this seems to be a typo.  The proposed changes replace the subscript $x$ with $d$ throughout the page where written in math notation or in code snippets, thus making the notation mathematically correct and consistent.